Player is damagable
Player has 6hp and the blob's attack now deals 1hp of damage. Currentley nothing changes besides a debug message when the player "dies".

Also fixed a bug that caused the blob to no longer chase the player. the issues lay in the interactions between the animator and physics. solved by setting update mode to use physics during the leap, and to operate normally when not leaping.